feat: add jwt token

+++ src/resolvers/project/mutation.ts
@@ -1,19 +1,36 @@
 import { Project } from '@typings';
+import { getUserIdFromAuthorizationHeader } from '../../utils/getUserIdFromAuthorizationHeader';
 
 export const Mutation = {
-  createProject: (_parent: unknown, args: Project.Mutation.CreateProject, { prisma }: Context): Promise<Project> => {
+  createProject: (_parent: unknown, args: Project.Mutation.CreateProject, { prisma, request }: Context): Promise<Project> => {
+    const id = getUserIdFromAuthorizationHeader(request);
     const owner = {
       connect: {
-        id: '', // TODO implement jwt token logic
+        id,
       },
     };
 
     return prisma.createProject({ ...args.data, owner });
   },
-  updateProject: (_parent: unknown, args: Project.Mutation.UpdateProject, { prisma }: Context): Promise<Project> => {
+  updateProject: async (_parent: unknown, args: Project.Mutation.UpdateProject, { prisma, request }: Context): Promise<Project> => {
+    const { where: { id } } = args;
+    const userId = getUserIdFromAuthorizationHeader(request);
+    const projectExists = await prisma.$exists.project({ id, owner: { id: userId } });
+
+    if (!projectExists) {
+      throw new Error('Project does not exist');
+    }
+
     return prisma.updateProject(args);
   },
-  deleteProject: (_parent: unknown, args: UniqueIdPayload, { prisma }: Context): Promise<Project> => {
+  deleteProject: async (_parent: unknown, args: UniqueIdPayload, { prisma, request }: Context): Promise<Project> => {
+    const { id } = args;
+    const userId = getUserIdFromAuthorizationHeader(request);
+    const projectExists = await prisma.$exists.project({ id, owner: { id: userId } });
+
+    if (!projectExists) {
+      throw new Error('Project does not exist');
+    }
     return prisma.deleteProject(args);
   },
 };
